// src/interfaces/GetUsers.interface.ts
export interface IUser {
  id: string;
  login: string;
  display_name: string;
  type: string;
  broadcaster_type: string;
  description: string;
  profile_image_url: string;
  offline_image_url: string;
  view_count: number;
  email: string;
};

export interface IUsersResponse {
  data: IUser[];
};